# api/swagger.yaml
openapi: 3.0.1
info:
  title: Emerald Web Api
  version: v1.0.0
servers:
- url: /
security:
- Bearer: []
paths:
...
  /api/Document/GetApplicationDocuments/{applicationId}:
    get:
      tags:
      - Document
      operationId: apiDocumentGetApplicationDocumentsApplicationIdGET
      parameters:
      - name: applicationId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: double
      responses:
        200:
          description: Success
      x-swagger-router-controller: Document
  /api/Document/LoanAgreement:
    post:
      tags:
      - Document
      operationId: apiDocumentLoanAgreementPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/RightSignatureRequest'
            example:
              ApplicationId: 2
              DocumentId: 33
          application/json:
            schema:
              $ref: '#/components/schemas/RightSignatureRequest'
            example:
              ApplicationId: 2
              DocumentId: 33
          text/json:
            schema:
              $ref: '#/components/schemas/RightSignatureRequest'
            example:
              ApplicationId: 2
              DocumentId: 33
          application/*+json:
            schema:
              $ref: '#/components/schemas/RightSignatureRequest'
            example:
              ApplicationId: 2
              DocumentId: 33
      responses:
        200:
          description: Success
      x-swagger-router-controller: Document
  /api/Document/Upload:
    post:
      tags:
      - Document
      operationId: apiDocumentUploadPOST
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/body'
            encoding:
              ApplicationId:
                style: form
              IsReceived:
                style: form
              DocumentType:
                style: form
              Document:
                style: form
      responses:
        200:
          description: Success
      x-swagger-router-controller: Document
...
  /api/Payment/ProcessPayment:
    post:
      tags:
      - Payment
      operationId: apiPaymentProcessPaymentPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        200:
          description: Success
      x-swagger-router-controller: Payment
  /api/Payment/PaymentResponse/{CompanyID}:
    get:
      tags:
      - Payment
      operationId: apiPaymentPaymentResponseCompanyIDGET
      parameters:
      - name: CompanyID
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: double
      responses:
        200:
          description: Success
      x-swagger-router-controller: Payment
...
  /api/Payment/Schedule/Generate:
    post:
      tags:
      - Payment
      operationId: apiPaymentScheduleGeneratePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleGenerateRequest'
            example:
              ScheduleType: 15
              StartDate: 04-Oct-2019
              Time: 12:00
              EndDate: 04-Nov-2019
              Amount: 50000
              PaymentMethod: 6
              IsAppend: false
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleGenerateRequest'
            example:
              ScheduleType: 15
              StartDate: 04-Oct-2019
              Time: 12:00
              EndDate: 04-Nov-2019
              Amount: 50000
              PaymentMethod: 6
              IsAppend: false
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
          text/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleGenerateRequest'
            example:
              ScheduleType: 15
              StartDate: 04-Oct-2019
              Time: 12:00
              EndDate: 04-Nov-2019
              Amount: 50000
              PaymentMethod: 6
              IsAppend: false
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
          application/*+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleGenerateRequest'
            example:
              ScheduleType: 15
              StartDate: 04-Oct-2019
              Time: 12:00
              EndDate: 04-Nov-2019
              Amount: 50000
              PaymentMethod: 6
              IsAppend: false
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
      responses:
        200:
          description: Success
      x-swagger-router-controller: Payment
  /api/Payment/Schedule/Create:
    post:
      tags:
      - Payment
      operationId: apiPaymentScheduleCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleCreateRequest'
            example:
              ApplicationID: 4448
              LoanID: 227
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleCreateRequest'
            example:
              ApplicationID: 4448
              LoanID: 227
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
          text/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleCreateRequest'
            example:
              ApplicationID: 4448
              LoanID: 227
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
          application/*+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleCreateRequest'
            example:
              ApplicationID: 4448
              LoanID: 227
              PaymentScheduleList:
              - DueDateStr: 04-Oct-2019
                DueDate: 2019-10-04T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 01-Oct-2019
                SubmitDate: 2019-10-01T12:00:00
              - DueDateStr: 18-Oct-2019
                DueDate: 2019-10-18T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 15-Oct-2019
                SubmitDate: 2019-10-15T12:00:00
              - DueDateStr: 01-Nov-2019
                DueDate: 2019-11-01T12:00:00
                Amount: 16666.666666666668
                PaymentMethod: 6
                PaymentMethodStr: GoCardless
                PaymentScheduleType: 15
                PaymentScheduleTypeStr: Bimonthly
                Time: 12:00
                SubmitDateStr: 29-Oct-2019
                SubmitDate: 2019-10-29T12:00:00
      responses:
        200:
          description: Success
      x-swagger-router-controller: Payment
  /api/Payment/ScheduleDetails:
    post:
      tags:
      - Payment
      operationId: apiPaymentScheduleDetailsPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
          text/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
          application/*+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
      responses:
        200:
          description: Success
      x-swagger-router-controller: Payment
  /api/Payment/ArrearsCleared:
    post:
      tags:
      - Payment
      operationId: apiPaymentArrearsClearedPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
          text/json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
          application/*+json:
            schema:
              $ref: '#/components/schemas/PaymentScheduleDetailRequest'
            example:
              ApplicationId: 99
              PageIndex: 1
              ItemsPage: 10
      responses:
        200:
          description: Success
      x-swagger-router-controller: Payment
components:
  schemas:
...
    RightSignatureRequest:
      type: object
      properties:
        applicationId:
          type: number
          format: double
        documentId:
          type: number
          format: double
      additionalProperties: false
...
    ProcessPaymentRequest:
      type: object
      properties:
        companyID:
          type: integer
          format: int32
        guaranterID:
          type: integer
          format: int32
        addressID:
          type: integer
          format: int32
      additionalProperties: false
    PaymentScheduleGenerateViewModel:
      type: object
      properties:
        dueDateStr:
          type: string
        dueDate:
          type: string
          format: date-time
        amount:
          type: number
          format: double
        paymentMethod:
          type: integer
          format: int32
        paymentMethodStr:
          type: string
        paymentScheduleType:
          type: integer
          format: int32
        paymentScheduleTypeStr:
          type: string
        time:
          type: string
        submitDateStr:
          type: string
        submitDate:
          type: string
          format: date-time
      additionalProperties: false
    PaymentScheduleGenerateRequest:
      type: object
      properties:
        scheduleType:
          type: integer
          format: int32
        startDate:
          type: string
        time:
          type: string
        endDate:
          type: string
        amount:
          type: number
          format: double
        paymentMethod:
          type: integer
          format: int32
        isAppend:
          type: boolean
        paymentScheduleList:
          type: array
          items:
            $ref: '#/components/schemas/PaymentScheduleGenerateViewModel'
      additionalProperties: false
    PaymentScheduleCreateRequest:
      type: object
      properties:
        applicationID:
          type: number
          format: double
        loanID:
          type: number
          format: double
        paymentScheduleList:
          type: array
          items:
            $ref: '#/components/schemas/PaymentScheduleGenerateViewModel'
      additionalProperties: false
    PaymentScheduleDetailRequest:
      type: object
      properties:
        applicationId:
          type: number
          format: double
        pageIndex:
          type: integer
          format: int32
        itemsPage:
          type: integer
          format: int32
      additionalProperties: false
    body:
      type: object
      properties:
        ApplicationId:
          type: number
          format: double
        IsReceived:
          type: boolean
        DocumentType:
          type: integer
          format: int32
        Document:
          type: string
          format: binary
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter\
        \ 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample:\
        \ \"Bearer 12345abcdef\""
      name: Authorization
      in: header